Sur/requirements.txt
django 
django-tailwind
gdal 
geos
postgresql
postgis
folium
django-leaflet
plotly
django-allauth
PyJWT
cryptography
django-allauth 
django-rest-auth
dj-rest-auth
django-sslserver
channels
channels-redis
daphne
